feat: add custom generator to generate only required models
