+++ test/index.js
@@ -59,7 +59,22 @@ const generateRealWorldSpecs = async () => {
 const main = async () => {
     await generate('./test/spec/v2.json', './test/generated/v2/');
     await generate('./test/spec/v3.json', './test/generated/v3/');
-    await OpenAPI.generateCustomSpec(generate, './test/spec/saddlebackApi.json', './test/generated/saddleback/');
+    await OpenAPI.generateCustomSpec({
+        input: './test/spec/saddlebackApi.json',
+        output: './test/generated/saddleback/',
+        httpClient: 'axios',
+        clientName: 'TestAxiosClassName',
+        useOptions: true,
+        useUnionTypes: false,
+        exportCore: true,
+        exportServices: true,
+        exportModels: true,
+        exportSchemas: false,
+        indent: '4',
+        postfix: '',
+        filterMethod: 'greedy',
+        filterArray: ['/api/agreement', '/api/agreement/{id}'],
+    });
     // await generateRealWorldSpecs();
 };
